Send a one-line notification for your own actions, so there's a record.
Using the hipchat notification, I found it hard to read because there
were messages responding to my own review request, but my own review
request wasn't listed in the chat.  So now I make it that I *do* get
messages for my own action, but only a short (one-line) summary.

I also change the colors around so that yellow is only used for
own-actions, and red ("action!") is used for new requests.