Update the README based on the fact we now use local versions of packages.
We get reviewboard from our local repository (so we can hack on it),
and likewise with python-hipchat, so the install instructions for them
have changed.  Update the README to represent the new world.

It was also confusing having install instructions spread between the
README and install.*, so I put most of them in the README now.

Reviewed-By: benkomalo,tom <http://reviewboard.khanacademy.org/r/60/>