Change to branch the release-1.6.x branch of reviewboard rather than HEAD.
The head version of reviewboard doesn't seem to be working for us.
I'm not interested in the new features, just making patches and the
like, so I don't have a problem with working on top of the latest
release.  (It may make life a bit more difficult if we decide to
upstream this stuff, or if we want to move to reviewboard 1.7, but
hopefully not.)