A script to help with manually updating the repository list for reviewboard.
Reviewboard needs to know what repositories we do code reviews
against.  We need to tell it manually.  While I hope to improve that
one day, or at least allow people to auto-register, for now I just
want to get a big list and do it by hand.

While it's too hard for me to do the actual registration
automatically: I'd need to deal with authentication and all the rest,
it's easy to emit urls that will pre-populate the form, so all I have
to do is paste the url into a web browser (while logged in as
reviewboard admin) and hit save.