Retry the git commands several times, only emitting text if they fail.
Without this, I was getting cron-mail for every transient failure.
Now it has to fail 3 times in a row before I get cron-mail.  The
cron-mail still only indicates there was a failure, though; we'll have
to go to th3e machien and look in cron.log to actually figure out what
the failure was.