Better retry fix: only print to stderr if the error is non-transient.
I changed the subprocess command to map stderr to stdout, so the two
are treated the same.  Instead, I print to stderr *only if* the
failure recurs even after the retries.  That way, callers can look at
stderr to determine if there was ever any error (they could also look
at the return code, but stderr gives a bit more information).

I use this for cron: stdout goes to the logfile, stderr gets sent as
email to the admin.