First steps toward installing and configuring phabricator (www.phabricator.org)
I added a config file needed for installation, and one (actually two,
but closely related) needed for running phabricator.

I also modified the README to include the phabricator install steps
I've done so far.  I reorganized the install step for the README to be
clearer; it was getting unwieldy with all the stuff I needed to
install for all the services running on this machine!