Clean up the lighttpd.conf, and coalesce all logs files into one directory.
I got rid of a lot of the reviewboard lighttpd.conf entry, which seems
mostly redundant (re-setting mime types) or actively harmful
(hijacking the lighttpd error logs, even for non-reviewboard uses).

As part of this, decided to try to put all the logs into one place.
We'll see how well that works.