Add a script that automatically adds new repositories to phabricator.
phabricator, like other code-review systems, needs to be told about
the repositories that we we are working on and might be sending
reviews against.

This script does that -- it knows where our repositories lives,
queries these places for the list of all repositories, and compares
that against the repositories phabricator already knows about.  If it
sees new ones, it adds them to the phabricator list.

The script also checks on the phabricator daemons to make sure they're
still all running ok.  (It needs to restart the daemons when new repos
are added, in any case.)

Phabricator has a full checkout of the repositories it manages, so
there's no need for us to keep a separate one.  So I've updated the
crontab to get rid of the mirrors we do; we'll rely on phabricator to
do the mirroring instead.  (Actually, I still enable the git mirror,
since it also serves as a git server.  Since it's fast, that's ok.)

I've updated the README to add the new instructions.  The major one is
making sure phabricator has the kiln tokens it needs to sync, and
moving the crontab setup to the end, so it doesn't start doing stuff
that might interfere with the setup work.