Add support for the --rr flag, including prompting for names.
Summary:
The functionality of --rr here matches that of our 'hg push' handler:
you can specify reviewers (either comma-separated, or multiple flags,
or both), and they will be passed along to the code review system to
start the review process.  As with the 'hg push' handler, the
reviewers can be specified by either username (phabricator username,
in this case) or real name, or a substring of either.  If the
specified reviewer is ambiguous, the script will prompt you to
disambiguate.

Note that this functionality is not currently useful, since arc does
not yet have support for the --reviewers commandline flag (though it's
been promised!)  But when --reviewers is pushed, then --rr will be
useful as well.

I also added a test for arc, since the functionality is getting more
complicated.

Reviewers: marcos, benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D65