Get rid of the check that all conduit responses start with a shield.
Summary:
It looks like the conduit API changed recently(?) to remove the
'for(;;)' that used to live at the beginning of each response.  That's
ok.  We still check if the shielf is there, and remove it if so, but
if it's not there, we keep going.

Also add in some more verbose-output that was useful for testing that
the fix was working as intended.

Test Plan: run update_phabricator_repos.py and verify it works again.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D656