Script to update phabricator on production.
Summary:
Updates the internal-webserver repository, which has the phabricator
source as 'fake sub-repos', and then pushes it to the phabricator
machine (toby), where it safely restarts the webserver there.

Script to update phabricator on production.

Updates the internal-webserver repository, which has the phabricator
source as 'fake sub-repos', and then pushes it to the phabricator
machine (toby), where it safely restarts the webserver there.

Oops, forgot to do the 'git pull' on production!

Check that the 'fake' sub-repos have been set up properly before running.

We could set them up for you if they're not, but since that's a
destructive operation, I think it's best to let the person run the
command themselves.  They only have to do it once.

Test Plan:
Ran 'bash -x update_phabricator.sh' and watched the output.  In
particular, watched that it properly restarted the daemons on the
production machine.

Ran 'bash -x update_phabricator.sh' and watched the output.  In
particular, watched that it properly restarted the daemons on the
production machine.

crossed fingers

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D940