Fix a bug around field labeling
As part of a previous review comment, I quickly snuck in a change to
strip chart label suffixes in a slick way.  However, the logic also
controlled which charts were returned and this caused an exception
to be thrown when unexpected chart labels (like
"Requests/Second (12 hrs)" rather than "Requests/Second") were returned.

I'm now sneaking in this back-out change.

Test plan:
I used a previously downloaded dashboard.html and ran this to verify
that the error was thrown before the fix went in and that the expected
results were printed after the fix:

<dashboard.html python dashboard_report.py --debug `date +%s`