We need to 'gc' bare repositories too.
Summary:
Without that, they get really unwieldy over time and a simple noop
'fetch' can take over a minute, when it would take about 2 seconds on
a new repo.

We don't need to gc all the time -- once a day would be more than
enough -- but there's no huge harm to doing it more often.  The --auto
flag will make the 'gc' a noop in the common case we don't fetch
anything from this repo.

Test Plan:
Ran
   git gc --auto --prune --aggressive --quiet
on khan-exercises.git on toby, which needed it badly.  It took about
10 minutes to run.  Ran it again and it returned immediately.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D1025