Fetch 100 repositories at a time to avoid rate-limiting.
Summary:
It turns out the reason this script was failing from time to time is
that github rate-limits us to 60 API calls an hour:
   http://developer.github.com/v3/#rate-limiting

While there are various ways around this, the easiest is to request
more results at a time, so we only need 1 (or 2) API calls per
script-run, rather than 4.  This should keep us going for a while,
anyway. :-)

Also add the per-page parameter to update_mirrors.sh

update_mirrors.sh still needs a lot of love -- it will start to miss
repos again once we get past 100 -- but this definitely helps, and is
low cost.

Test Plan:
Ran
   ./update_phabricator_repositories.py -n /var/tmp
and it said
   Adding 0 new repositories to phabricator
   Removing 0 deleted repositories from phabricator
showing the github call successfully got the right set of github repos.

Fingers crossed

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D3964