It turns out 'git add .' works just fine with fake repos, so use it.
Summary:
Once this is in, I'll update the forge page:
   https://sites.google.com/a/khanacademy.org/forge/for-developers/code-review-policy/using-phabricator/for-phabricator-admins
to remove the question about xargs.

Test Plan:
I ran
   ./update_phabricator.sh
with both the new code and the old code, each time aborting at the
'everything look ok' step.  I then compared the statuses that were
reported right before then, and they were functionally the same.  (The
new code puts more stuff in the staging index, but it will all come
out the same when the 'git commit -a' is run.)

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D4157