Phabricator has changed the way they do ssh keys.
Now you register an ssh username/key, which is stored in the
phabricator database, and you point repository.create() toward it.  We
have two ssh keys stored currently, one for kiln and one for git.

I'm not exactly sure if it's kosher to pass credentialPHID=None in the
case we don't need a credential; I guess we'll find out!  If not, I'll
have to make the repository.create() call a tiny bit uglier.

Auditors: alpert

Test plan:
Manually ran
   % python
   >>> import update_phabricator_repositories as u
   >>> phabricator_domain = 'http://phabricator.khanacademy.org'
   >>> phabctl = u.phabricator.Phabricator(host=phabricator_domain + '/api/')
   >>> q = phabctl.phid.lookup(names=['K1'])
   >>> q['K1']['phid']
   u'PHID-CDTL-...'

The rest is fingers-crossed; we'll see how it works once I deploy to
toby.