Give a better error message when we can't update the arcanist repo.
Summary: Make the error message even clearer when failing.

Test Plan:
Ran
   arc help
with these changes in, and it succeeded.

I didn't know how to get to a point where the git pull would give an
error, so it's fingers crossed on the actual quality of the error
message.

Fingers crossed.

Reviewers: alpert

Reviewed By: alpert

CC: laura

Differential Revision: http://phabricator.khanacademy.org/D5592