Yet another attempt to fix usernames with repos-to-clone.
The 'khanacademy@' username kept appearing and disappearing from the
repo-url as stored in phabricator, and I messed with the code to try
to handle it.  But I was really messing with the wrong code: I was
changing what we would pass to kiln via 'git clone'.  What I really
needed to be munging was the name as returned by phabricator.

Now I do that: I use the 'khanacademy@' username as part of the clone
url, as kiln requires, and when reading from phabricator, if I notice
the username is missing, I just add it in, and pretend phabricator
sent us the correct url-for-cloning all along.

Auditors: alpert

Test Plan:
Ran
   ./update_phabricator_repositories.py -n /tmp
with a new kiln repo that hadn't been added yet, and it just says:
   Adding 1 new repositories to phabricator
   Adding new repository Website/translations/crowdin: url=ssh://khanacademy@khanacademy.kilnhg.com/Website/translations/crowdin, callsign=GTC, vcs=git, destdir=/var/tmp/git/Website/translations/crowdin

So we're not re-adding existing ssh kiln-repos, which is good, and
we're getting the correct clone-url for this new repo, which is also
good.  I haven't actually tried making the phabricator API call that
adds this repo -- I'll do that on toby once this is checked in -- but
hopefully that will work too.