Cache the list of phabricator users rather than re-fetching all the time.
Summary:
The rule is we update the cache once a day (same time we update the
arc binary), and also anytime there's a cache miss -- that is, you
specify a reviewer-name that doesn't exist.  The latter is overly
conservative, perhaps, but hopefully will work well in practice.

Test Plan:
Ran
   arc help
   cat /tmp/arc.usernames
and saw output like:
   [["mgp", "Mike Parker"], ["alan", "Alan Pierce"], ["brianmerlob" ...

I then did
   arc diff --rr chris
and saw it work properly, with the timestamp of /tmp/arc.usernames
unchanged.

I then did
   arc diff --rr chrisss
and saw the timestamp of /tmp/arc.usernames update.

Reviewers: alpert, chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D7465