Update dashboard_report.py scraper to current App Engine UI
Summary:
There were 2 issues to fix:
1) Many reports changed names, although most values remained the same.
2) The default graph rendered on the page is now for the '24 hrs' period.

This isn't back to par, since it used to collect far more
data. Currently, only the "Summary" graph URL is contained in the
page. Before, all graphs for "6 hrs" were in the page. There appears to
be a new "/stats" endpoint that generates the graph URLs. In the future
we should hit that directly.

Test Plan:
  $ ./gae_dashboard_curl.py 'https://appengine.google.com/dashboard?app_id=s~khan-academy' chris@khanacademy.org >dashboard.html
  $ <dashboard.html ./dashboard_report.py -vn 0

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D8029