Don't tell phabricator to create webapp-foo repos.
HACK HACK HACK.  kiln is debugging some issue for us and cloning a
bunch of repos that we don't actually want/need phabricator to see.
And phabricator is keeling over trying to import them.

So I added a rule that we just ignore all repos that start with
'webapp' that aren't 'webapp' and 'webapp-backup'.

Auditors: alpert, chris

Test Plan:
Deleted webapp-debug and webapp-debug2 in phabricator, then ran
    ./update_phabricator_repositories.py -n /tmp
and it didn't suggest adding them back.  Nor did it suggest deleting
webapp or webapp-backup. :-)