Switch from using 'fake' sub-repos to using real ones.
Summary:
This fake sub-repo thing turned out to be a bit of a disaster: keeping
all the directories in sync was really hard, and having to check in
files twice (once in the 'fake' subrepo, once in the main repo) was
annoying.

We have lots of experience using 'normal' subrepos now, so change
internal-webserver to use that.

Test Plan:
Fingers crossed.  Once this is checked in I'll run
update_phabricator.sh and make sure it's still working right.

Reviewers: alpert, benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D8406