Change the script to require less set-up work by users.
We automatically figure out the upstream repo now, you don't need to
create it by hand when setting up the internal-webserver repo.

Likewise, when pushing, we automatically convert from http to ssh so
you don't have to enter your password all the time.  (We keep it as
ssh in the submodule definition because that's easier for toby, and
other clients that only pull and never push -- they don't have to set
up ssh keys with github.)

I did my best to make the summary info better now that we're using
real submodules instead of fake ones, but I can't figure out how to
get status.submodulesummary to do the right thing with respect to
merges.  So there's room for improvement there.

I also fixed what is probably a bug, where I didn't sync from origin
before syncing from upstream.  This wasn't a problem when only one
person was doing the syncing, but would be a problem if multiple
people were to do so.  But hopefully not anymore.

Auditors: alpert

Test Plan:
Ran ./update_phabricator.sh on my local machine and it seemed to work.