Update memcache and instance_summary parsers for missing fields
Summary:
When certain values can't be parsed from the input, I omit them from the
records that are sent to Graphite.

Test Plan:
/instance_summary returns the following in the browser:
1.9.5 714 total 0.725 364.1 ms 269.8 MBytes 1.9.6 1 total 0.000 Unknown ms 233.8 MBytes

And `./gae_dashboard/ka_report.py -A khan-academy -v -n` returns
something that looks about right for that data:

{'average_memory_mb': 272.0,
 'average_qps': 0.783,
 'num_instances': 675,
 'average_latency_ms': 390.3}

And is not crashing due to the "Unknown ms" field.

Fingers crossed on the /memcache oldest_item_age field.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D9768