Skip managed VMs when computing stats.
We only want stats for our webapp proper at the moment, not any
managed VMs we're playing around with in our khan-academy GAE app.  We
may change this in the future if we start using managed VMs for realz.

Pair programmed with chris.

Auditors: chris

Test Plan:
Fingers crossed.  We'll see if this quiets the email spam!