A script to collect specific bigquery data and email it out.
Summary:
The idea is we'll run this in a cron job daily so we get some nice
info about how our site is doing: for instance, what routes are
causing the most Out Of Memory errors, etc.

(Perhaps we should put the raw data somewhere else other than in an
email, but we'll do this for now and see how useful it is!)

Test Plan:
Fingers crossed.  I don't have sendmail installed locally so had a
hard time testing it.  Will run from toby once this is checked in.

Reviewers: chris

Reviewed By: chris

Subscribers: benkomalo, marcia

Differential Revision: http://phabricator.khanacademy.org/D11365