Adds util for embedding images in emails
Summary:
This is the first piece of adding sparkline graphs to the email reports.
Unfortunately, Gmail won't display embedded SVG (and certainly not canvas/js),
so we will have to render the graphs to PNG, and embed those.

Test Plan:
Tested in a shell, putting images into an HTML body and sending it to myself.
The images display correctly in Gmail.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: benkomalo, marcia, alan, chris

Differential Revision: http://phabricator.khanacademy.org/D11605