Collect several stats per-module.
Summary:
This involved adding a routine to list all the GAE modules, then
changing some stuff that was only done once to be done in a 'for
module in modules:' loop.  I also had to pass through the module name
until (eventually) the send-to-graphite routine, which now uses the
module-name as part of the key when appropriate.

Test Plan:
   env srcdir=. ./fetch_stats.sh -n -v
gives reasonable-looking output, which this editor window is too small
to contain.

Reviewers: chris

Reviewed By: chris

Subscribers: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D13241