Change the perf charts to use the new URL route format
Summary:
Since I changed the url_route format in D13968 , we need to update the perf
graph code accordingly.

I made this change by going to the code for each of the routes that needed to be
changed, and I copied the string from the @route declaration and used it here.
If D13968 was written correctly, this should be the new format.

Test Plan:
Fingers crossed. If I made a mistake, it'll be obvious (some url route will drop
to 0 requests) and we can fix it.

Reviewers: jlfwong

Subscribers: csilvers, benkomalo

Differential Revision: http://phabricator.khanacademy.org/D13969