Escape HTML characters in daily BQ emails
Summary:
Now that URL routes can look like "api.main:/api/internal/foo/<bar>", we need to
be careful that we HTML-escape the strings so they look right in the emails.

Test Plan:
Fingers crossed. My understanding is that this code is hard to test, and a
failure isn't that big of a deal anyway.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D14113