Add the task computation route to the performance tracking page
Summary:
This route returns the list of learning tasks to be rendered for the student on that mission's learning dashboard. It is sensitive to changes to the knowledge model parameters because it must compute a lot of predictions to determine the recommendations to make.

I found this elog_url_route with this bigquery:

```
select elog_url_route
from [logs.requestlogs_20150124]
where elog_url_route contains '/descriptors'
```

This is related to the changes in D15871.

Test Plan: Once this is approved and deployed I will look at perf.khanacademy.org. I don't plan on doing any other testing, since this is a trivial change.

Reviewers: jlfwong

Reviewed By: jlfwong

Subscribers: benkomalo, jace

Differential Revision: https://phabricator.khanacademy.org/D15893