Fix URL route on perf dashboard
Summary:
I update the name from user_data_key to kaid, so now I should change the URL
route name as well. This was the only route with a user_data_key on the
dashboard, so this should be the only one that needs fixing for the user ID
reform project.

Test Plan: Fingers crossed.

Reviewers: jlfwong, csilvers

Reviewed By: csilvers

Subscribers: csilvers, benkomalo

Differential Revision: https://phabricator.khanacademy.org/D16434