Fix the SQL query to match the new logs output as of D16705.
Summary:
Now we log counts rather than a bunch of times, making the sql
simpler!  I model the new sql on the existing 'cost' field, which is
already a count.

I also fixed a bug in the client-api-usage report.  Previously, it was
putting everything into the 'Web Browsers/other' bucket.  Now it
should start putting some stuff in the ios bucket again.

Depends on: D16705

Test Plan:
Fingers crossed!  If I got this wrong, all that will happen is we
won't get our daily stats email.  I'll notice that and fix it.

Reviewers: chris, alan

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D16707