Update _MODULE_CPU_COUNT with more accurate utilization info.
I ran the bigquery query to calculate the new numbers.  We now take
into account the fact that the single-threaded modules do not have
perfect utilization; there's some idle time built in due to
overcapacity, to be able to handle traffic spikes.

This will make it so the costs of routes are higher, which will more
accurately match their cost to us.  The ratios between modules are
more or less the same, though multithreaded modules are now about 3x
as utilized as single-threaded, not 4.

Auditors: alan, chris

Test plan:
None