Add script to export graphite metrics to Cloud Monitoring
Summary:
Khan Academy has all sorts of time-series data in HostedGraphite
describing deployments, user events on the production website, server
health, error counts, billed resources, and so on.

Google's Cloud Monitoring integrates nicely with App Engine, which
runs www.khanacademy.org, and has some very nice alerting features to
notify developers via HipChat, PagerDuty, email, &c. when things go
wrong.

This script is the bridge that enables Cloud Monitoring to trigger
alerts based on time-series data in HostedGraphite. It's expected that
a cron job will periodically run this script.

This changeset also requires updating the setup scripts in the
aws-config repo to have the right secrets in the right places. That'll
be a separate commit.

Test Plan:
I ran a few test modes of the script:

 ./graphite_bridge.py -vn
 ./graphite_bridge.py -vt

And then I ran it for real, and verified the results show up in Cloud
Monitoring:

 ./graphite_bridge.py

Reviewers: jlfwong, csilvers

Reviewed By: csilvers

Subscribers: benkomalo, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D17298