Bug-fixes for aggregate and historical data
Summary:
In practice, for conversion data, it makes more sense to look at 5
minute buckets than 5 second buckets. The value tends to be more stable
and so is better for avoiding spurious alerts.

The first bug fix is to skip the first bucket of data. This bucket can
be considered "active". It's currently filling with data so it's value
isn't very stable. For more stable data, we throw it out and look at a
bucket whose time is passed and whose value is unchanging.

The second bug fix is to skip values of 1.0 for my week-over-week ratio
values. For the ":sum" values, HostedGraphite's most recent buckets tend
to be filled with null values. The data is about 15 minutes behind the
current time. Null values are fine, and I skip them. However, when
they're used in a compound target, like diffSeries(seriesA, seriesB), a
null value throws off the calculation. So when we're trying to calculate the
current difference from historical data, we'll always get:

  (0 - historical) / historical, or 1.0.   (Not -1.0, oddly enough, and not because of absolute()).

A solution is to use keepLastValue to ensure that there aren't any
nulls. The downside is that a bucket's value might be inaccurate if it
hasn't been filled and is being reported as an older value.

Test Plan:
I ran `./graphite_bridge.py -vn` and saw reasonable output. Before the
change, the week_over_week values for bingo things were all 1.0:

  INFO:root:Sending bingo.login.week_over_week 0.0595300261097 1430159700
  INFO:root:Sending bingo.registration.week_over_week 0.04 1430159700
  INFO:root:Sending bingo.video_started.week_over_week 0.0959160734357 1430159700
  INFO:root:Sending batch_module.average_latency_ms.week_over_week 0.5768568244 1430159700
  INFO:root:Sending default_module.average_latency_ms 262.0 1430159700
  INFO:root:Sending bingo.problem_attempt.week_over_week 0.172362921072 1430159700
  INFO:root:Sending batch_module.average_latency_ms 417.6 1430159700
  Would send 7 datapoint(s)

Reviewers: alan, csilvers

Reviewed By: csilvers

Subscribers: csilvers, benkomalo, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D17534