Move fetch_stats.py over to using the cloudmonitoring API.
Summary:
The old technique of scraping the dashboard has gone away, instead we
can use a spiffy API.  This commit adds the infrastructure needed to
fetch the data from the API and send it to hipchat, and implements the
code for one of the stats.

I made some boilerplate for each of the stats that we currently
support in graphite.  We probably won't be able to re-create those
stats entirely using the new API; as I go through the stats I'll
figure out which ones we can keep and which we can't.

This doesn't attempt to fold ka_report.py into fetch_stats.py, though
we'll probably want to do that eventually (if it's possible).

Part of the fix for https://app.asana.com/0/27216215224639/37153007969091

Properly convert the stats to be per-second.

I was forgetting to do that conversion before!

I also implemented another easy summary.

Add a few more stats, and factor out some convenience methods.

Now the functions are starting to look a lot cleaner!  Hopefully it's
not so many levels of functions between the high-level function and
the actual work that it's hard to figure out what's going on.

Add a bunch more stats, that were easy to add

Get rid of some stats we can't compute and don't really care about.

Aggregate memory use isn't really that useful.  memcache operations
aren't either.

We've never used the per-instance memcache stats.

memcache_size_mb is already covered by another stat
(memcache.total_cache_size_bytes).  And all the usage.* stats are
really covered by fetch_usage.py.

Write the last bits of the project, to automatically run over time.

We now maintain the last-time-read in a simple file db, and update it
when we process.

Fixed a bug in how we called the cloudmonitoring API.

I was switching youngest and oldest!  Thanks chris for catching that.

I also got rid of _send_summary_to_graphite() which was a confusing
name.  I think this new version reads pretty well anyway, and may be
easier to reason about for people new to the code.

Test Plan:
I installed the cloudmonitoring secret into my homedir by copying it
from toby, then ran
   env PATH=$HOME/khan/webapp/third_party/frankenserver:$PATH ./fetch_stats.py -n

It emitted:
  --> Would send to graphite:
  (u'w.g.d.s.default_module.dos_api_denials_per_second', (1438099755, 0))
  (u'w.g.d.s.default_module.dos_api_denials_per_second', (1438099815, 0))
  (u'w.g.d.s.default_module.dos_api_denials_per_second', (1438099875, 0))
  (u'w.g.d.s.default_module.dos_api_denials_per_second', (1438099935, 0))
  (u'w.g.d.s.default_module.dos_api_denials_per_second', (1438099995, 0))
  --> Would send to graphite:
  (u'w.g.d.s.frontend_highmem_module.dos_api_denials_per_second', (1438099755, 0))
  [etc]
which looks right to me.

(same)

Same, but I changed the stat we're collecting to get real data.  Now
the output looks like:

  --> Would send to graphite:
  (u'w.g.d.s.default_module.client_errors_per_second', (1438099714, 3.7166666666666663))
  (u'w.g.d.s.default_module.client_errors_per_second', (1438099774, 4.483333333333333))
  (u'w.g.d.s.default_module.client_errors_per_second', (1438099834, 4.5))
  (u'w.g.d.s.default_module.client_errors_per_second', (1438099894, 3.8))
  (u'w.g.d.s.default_module.client_errors_per_second', (1438099954, 3.6166666666666667))
  [...]
  --> Would send to graphite:
  ('w.g.d.s.client_errors_per_second', (1438099714, 4.3166666666666655))
  ('w.g.d.s.client_errors_per_second', (1438099774, 4.916666666666666))
  ('w.g.d.s.client_errors_per_second', (1438099834, 5.233333333333333))
  ('w.g.d.s.client_errors_per_second', (1438099894, 4.366666666666666))
  ('w.g.d.s.client_errors_per_second', (1438099954, 3.9))

None new

(None new)

Ran
   env PATH=$HOME/khan/webapp/third_party/frankenserver:$PATH ./fetch_stats.py -nv
on my local machine, and saw it print a bunch of stuff, including:
   Calculating stats for memcache.total_cache_size_bytes from 1435541159 - 1435541459
   [...]
   Would update last-processed-time from 1438130164 to 1438131064
   Done!

Reviewers: chris

Reviewed By: chris

Subscribers: benkomalo, davidwang, ethan

Differential Revision: https://phabricator.khanacademy.org/D19934