Write a new version of fetch_usage that gets the data from the csv file.
Summary:
Chris already flipped thd switch to save the csv data to gcs every
night.  This can read that data and make a bunch of graphite stats out
of it.

Fix a typo involving where I did int().

Interestingly, the old code did the right thing, since in practice
tz_offset[4:] is always '00', and int('000000000...00000000') is still
0.

Finish up this code by adding the run-over-time functionality.

I also made the code robust to 'Credit1 Amount' missing, since it
seems to be a new field introduced in the past month.

Test Plan:
Ran
   ./fetch_stats.py
and saw the output like this:
    --> Would send to graphite:
    ('webapp.gcp.khan_academy.usage.appengine.backend_instances.seconds', (1437980399, 4009663.23413))
    ('webapp.gcp.khan_academy.usage.appengine.backend_instances.cost', (1437980399, 55.689767))
    ('webapp.gcp.khan_academy.usage.appengine.backend_instances.credit', (1437980399, 0.0))
    ('webapp.gcp.khan_academy.usage.appengine.blobstore_storage.byte_seconds', (1437980399, 26200346374771200))
    ('webapp.gcp.khan_academy.usage.appengine.blobstore_storage.cost', (1437980399, 0.236867))
    ('webapp.gcp.khan_academy.usage.appengine.blobstore_storage.credit', (1437980399, 0.0))
    ...
    ('webapp.gcp.khan_academy.usage.pubsub.message_operations.requests', (1438066799, 22322064))
    ('webapp.gcp.khan_academy.usage.pubsub.message_operations.cost', (1438066799, 8.928771))
    ('webapp.gcp.khan_academy.usage.pubsub.message_operations.credit', (1438066799, 0.0))

which seems plausible.

Ran ./fetch_usage.py -n
and had it print a bunch of stuff out.

Reviewers: chris

Reviewed By: chris

Subscribers: benkomalo, davidwang, ethan

Differential Revision: https://phabricator.khanacademy.org/D19952